# remote_openapi.json
{"openapi":"3.1.0","info":{"title":"Orbis Ethica API","version":"2.0.0"},"paths":{"/api/verify":{"post":{"summary":"Verify Action","description":"Verify an action's ethical alignment.\nRequires 0.1 ETHC payment.","operationId":"verify_action_api_verify_post","parameters":[{"name":"X-Orbis-Wallet","in":"header","required":false,"schema":{"type":"string","title":"X-Orbis-Wallet"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DecisionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/balance/{wallet_id}":{"get":{"summary":"Get Balance","description":"Get the ETHC balance for a wallet.","operationId":"get_balance_api_balance__wallet_id__get","parameters":[{"name":"wallet_id","in":"path","required":true,"schema":{"type":"string","title":"Wallet Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/{wallet_id}":{"get":{"summary":"Get History","description":"Get transaction history for a wallet.","operationId":"get_history_api_history__wallet_id__get","parameters":[{"name":"wallet_id","in":"path","required":true,"schema":{"type":"string","title":"Wallet Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/stats":{"get":{"summary":"Get Stats","description":"Get real-time system statistics from the DB.","operationId":"get_stats_api_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DecisionResponse":{"properties":{"safe":{"type":"boolean","title":"Safe"},"reason":{"type":"string","title":"Reason"},"risk_level":{"type":"string","title":"Risk Level"},"score":{"type":"number","title":"Score"},"breakdown":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Breakdown"}},"type":"object","required":["safe","reason","risk_level","score","breakdown"],"title":"DecisionResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyRequest":{"properties":{"action":{"type":"string","title":"Action"},"context":{"additionalProperties":true,"type":"object","title":"Context","default":{}}},"type":"object","required":["action"],"title":"VerifyRequest"}}}}